Network.HTTP.Client.Internal

Description

Note that this is essentially the "kitchen sink" export module, including many functions intended only to be used internally by this package. No API stability is guaranteed for this module. If you see functions here which you believe should be promoted to a stable API, please contact the author.

Low-level response body handling

makeChunkedReader

Arguments

:: Bool
raw

-> Connection

-> IO BodyReader

makeLengthReader :: Int -> Connection -> IO BodyReader

makeGzipReader :: BodyReader -> IO BodyReader

makeUnlimitedReader :: Connection -> IO BodyReader

brConsume :: BodyReader -> IO [ByteString]

Strictly consume all remaining chunks of data from the stream.

Since 0.1.0

brEmpty :: BodyReader

brAddCleanup :: IO () -> BodyReader -> BodyReader

brReadSome :: BodyReader -> Int -> IO ByteString

Continuously call brRead, building up a lazy ByteString until a chunk is constructed that is at least as many bytes as requested.

Since 0.4.20

brRead :: BodyReader -> IO ByteString

Get a single chunk of data from the response body, or an empty bytestring if no more data is available.

Note that in order to consume the entire request body, you will need to repeatedly call this function until you receive an empty ByteString as a result.

Since 0.1.0

Raw connection handling

connectionReadLine :: Connection -> IO ByteString

connectionReadLineWith :: Connection -> ByteString -> IO ByteString

connectionDropTillBlankLine :: Connection -> IO ()

Keep dropping input until a blank line is found.

dummyConnection

Arguments

:: [ByteString]
input

-> IO (Connection, IO [ByteString], IO [ByteString])
conn, output, input

For testing

openSocketConnection

Arguments

:: (Socket -> IO ())

-> Maybe HostAddress

-> String
host

-> Int
port

-> IO Connection

openSocketConnectionSize

Arguments

:: (Socket -> IO ())

-> Int
chunk size

-> Maybe HostAddress

-> String
host

-> Int
port

-> IO Connection

makeConnection

Arguments

:: IO ByteString
read

-> (ByteString -> IO ())
write

-> IO ()
close

-> IO Connection

Cookies

updateCookieJar

Arguments

:: Response a
Response received from server

-> Request
Request which generated the response

-> UTCTime
Value that should be used as "now"

-> CookieJar
Current cookie jar

-> (CookieJar, Response a)
(Updated cookie jar with cookies from the Response, The response stripped of any "Set-Cookie" header)

This applies receiveSetCookie to a given Response

receiveSetCookie

Arguments

:: SetCookie
The SetCookie the cookie jar is receiving

-> Request
The request that originated the response that yielded the SetCookie

-> UTCTime
Value that should be used as "now"

-> Bool
Whether or not this request is coming from an "http" source (not javascript or anything like that)

-> CookieJar
Input cookie jar to modify

-> CookieJar
Updated cookie jar

This corresponds to the algorithm described in Section 5.3 "Storage Model" This function consists of calling generateCookie followed by insertCheckedCookie. Use this function if you plan to do both in a row. generateCookie and insertCheckedCookie are only provided for more fine-grained control.

generateCookie

Arguments

:: SetCookie
The SetCookie we are encountering

-> Request
The request that originated the response that yielded the SetCookie

-> UTCTime
Value that should be used as "now"

-> Bool
Whether or not this request is coming from an "http" source (not javascript or anything like that)

-> Maybe Cookie
The optional output cookie

Turn a SetCookie into a Cookie, if it is valid

insertCheckedCookie

Arguments

:: Cookie
The SetCookie the cookie jar is receiving

-> CookieJar
Input cookie jar to modify

-> Bool
Whether or not this request is coming from an "http" source (not javascript or anything like that)

-> CookieJar
Updated (or not) cookie jar

Insert a cookie created by generateCookie into the cookie jar (or not if it shouldn't be allowed in)

insertCookiesIntoRequest

Arguments

:: Request
The request to insert into

-> CookieJar
Current cookie jar

-> UTCTime
Value that should be used as "now"

-> (Request, CookieJar)
(Ouptut request, Updated cookie jar (last-access-time is updated))

This applies the computeCookieString to a given Request

computeCookieString

Arguments

:: Request
Input request

-> CookieJar
Current cookie jar

-> UTCTime
Value that should be used as "now"

-> Bool
Whether or not this request is coming from an "http" source (not javascript or anything like that)

-> (ByteString, CookieJar)
(Contents of a "Cookie" header, Updated cookie jar (last-access-time is updated))

This corresponds to the algorithm described in Section 5.4 "The Cookie Header"

evictExpiredCookies

Arguments

:: CookieJar
Input cookie jar

-> UTCTime
Value that should be used as "now"

-> CookieJar
Filtered cookie jar

This corresponds to the eviction algorithm described in Section 5.3 "Storage Model"

createCookieJar :: [Cookie] -> CookieJar

destroyCookieJar :: CookieJar -> [Cookie]

pathMatches :: ByteString -> ByteString -> Bool

This corresponds to the subcomponent algorithm entitled "Path-Match" detailed in section 5.1.4

removeExistingCookieFromCookieJar :: Cookie -> CookieJar -> (Maybe Cookie, CookieJar)

domainMatches

Arguments

:: ByteString
Domain to test

-> ByteString
Domain from a cookie

-> Bool

This corresponds to the subcomponent algorithm entitled "Domain Matching" detailed in section 5.1.3

isIpAddress :: ByteString -> Bool

defaultPath :: Request -> ByteString

This corresponds to the subcomponent algorithm entitled "Paths" detailed in section 5.1.4

Performing requests

withResponse :: Request -> Manager -> (Response BodyReader -> IO a) -> IO a

Perform a Request using a connection acquired from the given Manager, and then provide the Response to the given function. This function is fully exception safe, guaranteeing that the response will be closed when the inner function exits. It is defined as:

withResponse req man f = bracket (responseOpen req man) responseClose f

It is recommended that you use this function in place of explicit calls to responseOpen and responseClose.

You will need to use functions such as brRead to consume the response body.

Since 0.1.0

httpLbs :: Request -> Manager -> IO (Response ByteString)

A convenience wrapper around withResponse which reads in the entire response body and immediately closes the connection. Note that this function performs fully strict I/O, and only uses a lazy ByteString in its response for memory efficiency. If you are anticipating a large response body, you are encouraged to use withResponse and brRead instead.

Since 0.1.0

httpNoBody :: Request -> Manager -> IO (Response ())

A convenient wrapper around withResponse which ignores the response body. This is useful, for example, when performing a HEAD request.

Since 0.3.2

httpRaw :: Request -> Manager -> IO (Response BodyReader)

Get a Response without any redirect following.

httpRaw' :: Request -> Manager -> IO (Request, Response BodyReader)

Get a Response without any redirect following.

This extended version of httpRaw also returns the potentially modified Request.

getModifiedRequestManager :: Manager -> Request -> IO (Manager, Request)

The used Manager can be overridden (by requestManagerOverride) and the used Request can be modified (through managerModifyRequest). This function allows to retrieve the possibly overridden Manager and the possibly modified Request.

(In case the Manager is overridden by requestManagerOverride, the Request is being modified by managerModifyRequest of the new Manager, not the old one.)

responseOpen :: Request -> Manager -> IO (Response BodyReader)

The most low-level function for initiating an HTTP request.

The first argument to this function gives a full specification on the request: the host to connect to, whether to use SSL, headers, etc. Please see Request for full details. The second argument specifies which Manager should be used.

This function then returns a Response with a BodyReader. The Response contains the status code and headers that were sent back to us, and the BodyReader contains the body of the request. Note that this BodyReader allows you to have fully interleaved IO actions during your HTTP download, making it possible to download very large responses in constant memory.

An important note: the response body returned by this function represents a live HTTP connection. As such, if you do not use the response body, an open socket will be retained indefinitely. You must be certain to call responseClose on this response to free up resources.

This function automatically performs any necessary redirects, as specified by the redirectCount setting.

When implementing a (reverse) proxy using this function or relating functions, it's wise to remove Transfer-Encoding:, Content-Length:, Content-Encoding: and Accept-Encoding: from request and response headers to be relayed.

Since 0.1.0

responseClose :: Response a -> IO ()

Close any open resources associated with the given Response. In general, this will either close an active Connection or return it to the Manager to be reused.

Since 0.1.0

applyCheckStatus :: Request -> (Status -> ResponseHeaders -> CookieJar -> Maybe SomeException) -> Response BodyReader -> IO (Maybe SomeException)

Apply 'Request'\'s checkStatus and return resulting exception if any.

httpRedirect

Arguments

:: Int
redirectCount

-> (Request -> IO (Response BodyReader, Maybe Request))
function which performs a request and returns a response, and possibly another request if there's a redirect.

-> Request

-> IO (Response BodyReader)

Redirect loop.

httpRedirect'

Arguments

:: Int
redirectCount

-> (Request -> IO (Response BodyReader, Request, Bool))
function which performs a request and returns a response, the potentially modified request, and a Bool indicating if there was a redirect.

-> Request

-> IO (Request, Response BodyReader)

Redirect loop.

This extended version of httpRaw also returns the Request potentially modified by managerModifyRequest.

Parse response headers

parseStatusHeaders :: Connection -> Maybe Int -> Maybe (IO ()) -> IO StatusHeaders

Request helper functions

parseUrl :: MonadThrow m => String -> m Request

Deprecated: Please use parseUrlThrow, parseRequest, or parseRequest_ instead

Deprecated synonym for parseUrlThrow. You probably want parseRequest or parseRequest_ instead.

Since: 0.1.0

parseUrlThrow :: MonadThrow m => String -> m Request

Same as parseRequest, except will throw an HttpException in the event of a non-2XX response.

Since: 0.4.30

parseRequest :: MonadThrow m => String -> m Request

Convert a URL into a Request.

This defaults some of the values in Request, such as setting method to GET and requestHeaders to [].

Since this function uses MonadThrow, the return monad can be anything that is an instance of MonadThrow, such as IO or Maybe.

You can place the request method at the beginning of the URL separated by a space, e.g.:

@@ parseRequeset "POST http://httpbin.org/post" @@

Note that the request method must be provided as all capital letters.

Since: 0.4.30

parseRequest_ :: String -> Request

Same as parseRequest, but in the cases of a parse error generates an impure exception. Mostly useful for static strings which are known to be correctly formatted.

defaultRequest :: Request

A default request value

Since: 0.4.30

setUriRelative :: MonadThrow m => Request -> URI -> m Request

Add a URI to the request. If it is absolute (includes a host name), add it as per setUri; if it is relative, merge it with the existing request.

getUri :: Request -> URI

Extract a URI from the request.

Since 0.1.0

setUri :: MonadThrow m => Request -> URI -> m Request

Validate a URI, then add it to the request.

browserDecompress :: ByteString -> Bool

Decompress a compressed stream unless the content-type is 'application/x-tar'.

alwaysDecompress :: ByteString -> Bool

Always decompress a compressed stream.

addProxy :: ByteString -> Int -> Request -> Request

Add a proxy to the Request so that the Request when executed will use the provided proxy.

Since 0.1.0

applyBasicAuth :: ByteString -> ByteString -> Request -> Request

Add a Basic Auth header (with the specified user name and password) to the given Request. Ignore error handling:

 applyBasicAuth "user" "pass" $ parseRequest_ url

Since 0.1.0

applyBasicProxyAuth :: ByteString -> ByteString -> Request -> Request

Add a Proxy-Authorization header (with the specified username and password) to the given Request. Ignore error handling:

applyBasicProxyAuth "user" "pass" <$> parseRequest "http://example.org"

Since 0.3.4

urlEncodedBody :: [(ByteString, ByteString)] -> Request -> Request

Add url-encoded parameters to the Request.

This sets a new requestBody, adds a content-type request header and changes the method to POST.

Since 0.1.0

needsGunzip

Arguments

:: Request

-> [Header]
response headers

-> Bool

requestBuilder :: Request -> Connection -> IO (Maybe (IO ()))

useDefaultTimeout :: Maybe Int

Magic value to be placed in a Request to indicate that we should use the timeout value in the Manager.

Since 1.9.3

setRequestIgnoreStatus :: Request -> Request

Modify the request so that non-2XX status codes do not generate a runtime StatusCodeException.

Since: 0.4.29

setQueryString :: [(ByteString, Maybe ByteString)] -> Request -> Request

Set the query string to the given key/value pairs.

Since 0.3.6

streamFile :: FilePath -> IO RequestBody

Send a file as the request body.

It is expected that the file size does not change between calling streamFile and making any requests using this request body.

Since 0.4.9

observedStreamFile :: (StreamFileStatus -> IO ()) -> FilePath -> IO RequestBody

Send a file as the request body, while observing streaming progress via a PopObserver. Observations are made between reading and sending a chunk.

It is expected that the file size does not change between calling observedStreamFile and making any requests using this request body.

Since 0.4.9

extractBasicAuthInfo :: URI -> Maybe (ByteString, ByteString)

Extract basic access authentication info in URI. Return Nothing when there is no auth info in URI.

Low-level response body handling

getRedirectedRequest :: Request -> ResponseHeaders -> CookieJar -> Int -> Maybe Request

If a request is a redirection (status code 3xx) this function will create a new request from the old request, the server headers returned with the redirection, and the redirection code itself. This function returns Nothing if the code is not a 3xx, there is no location header included, or if the redirected response couldn't be parsed with parseRequest.

If a user of this library wants to know the url chain that results from a specific request, that user has to re-implement the redirect-following logic themselves. An example of that might look like this:

myHttp req man = do
   (res, redirectRequests) <- (`runStateT` []) $
        'httpRedirect'
            9000
            (\req' -> do
               res <- http req'{redirectCount=0} man
               modify (\rqs -> req' : rqs)
               return (res, getRedirectedRequest req' (responseHeaders res) (responseCookieJar res) (W.statusCode (responseStatus res))
               )
            'lift'
            req
   applyCheckStatus (checkStatus req) res
   return redirectRequests

getResponse

Arguments

:: ConnRelease

-> Maybe Int

-> Request

-> Connection

-> Maybe (IO ())
Action to run in case of a '100 Continue'.

-> IO (Response BodyReader)

lbsResponse :: Response BodyReader -> IO (Response ByteString)

Convert a Response that has a Source body to one with a lazy ByteString body.

Manager

data ManagerSettings

Settings for a Manager. Please use the defaultManagerSettings function and then modify individual settings. For more information, see http://www.yesodweb.com/book/settings-types.

Since 0.1.0

Constructors

ManagerSettings

Fields

managerConnCount :: Int

Number of connections to a single host to keep alive. Default: 10.

Since 0.1.0

managerRawConnection :: IO (Maybe HostAddress -> String -> Int -> IO Connection)

Create an insecure connection.

Since 0.1.0 FIXME in the future, combine managerTlsConnection and managerTlsProxyConnection

managerTlsConnection :: IO (Maybe HostAddress -> String -> Int -> IO Connection)

Create a TLS connection. Default behavior: throw an exception that TLS is not supported.

Since 0.1.0

managerTlsProxyConnection :: IO (ByteString -> (Connection -> IO ()) -> String -> Maybe HostAddress -> String -> Int -> IO Connection)

Create a TLS proxy connection. Default behavior: throw an exception that TLS is not supported.

Since 0.2.2

managerResponseTimeout :: Maybe Int

Default timeout (in microseconds) to be applied to requests which do not provide a timeout value.

Default is 30 seconds

Since 0.1.0

managerRetryableException :: SomeException -> Bool

Exceptions for which we should retry our request if we were reusing an already open connection. In the case of IOExceptions, for example, we assume that the connection was closed on the server and therefore open a new one.

Since 0.1.0

managerWrapIOException :: forall a. IO a -> IO a

Action wrapped around all attempted Requests, usually used to wrap up exceptions in library-specific types.

Default: wrap all IOExceptions in the InternalIOException constructor.

Since 0.1.0

managerIdleConnectionCount :: Int

Total number of idle connection to keep open at a given time.

This limit helps deal with the case where you are making a large number of connections to different hosts. Without this limit, you could run out of file descriptors.

Default: 512

Since 0.3.7

managerModifyRequest :: Request -> IO Request

Perform the given modification to a Request before performing it.

Default: no modification

Since 0.4.4

managerProxyInsecure :: ProxyOverride

How HTTP proxy server settings should be discovered.

Default: respect the proxy value on the Request itself.

Since 0.4.7

managerProxySecure :: ProxyOverride

How HTTPS proxy server settings should be discovered.

Default: respect the proxy value on the Request itself.

Since 0.4.7

newManager :: ManagerSettings -> IO Manager

Create a Manager. The Manager will be shut down automatically via garbage collection.

Creating a new Manager is a relatively expensive operation, you are advised to share a single Manager between requests instead.

The first argument to this function is often defaultManagerSettings, though add-on libraries may provide a recommended replacement.

Since 0.1.0

closeManager :: Manager -> IO ()

Deprecated: Manager will be closed for you automatically when no longer in use

Close all connections in a Manager.

Note that this doesn't affect currently in-flight connections, meaning you can safely use it without hurting any queries you may have concurrently running.

Since 0.1.0

withManager :: ManagerSettings -> (Manager -> IO a) -> IO a

Deprecated: Use newManager instead

Create, use and close a Manager.

Since 0.2.1

getConn :: Request -> Manager -> IO (ConnRelease, Connection, ManagedConn)

failedConnectionException :: Request -> HttpException

Create an exception to be thrown if the connection for the given request fails.

defaultManagerSettings :: ManagerSettings

Default value for ManagerSettings.

Note that this value does not have support for SSL/TLS. If you need to make any https connections, please use the http-client-tls package, which provides a tlsManagerSettings value.

Since 0.1.0

rawConnectionModifySocket :: (Socket -> IO ()) -> IO (Maybe HostAddress -> String -> Int -> IO Connection)

A value for the managerRawConnection setting, but also allows you to modify the underlying Socket to set additional settings. For a motivating use case, see: https://github.com/snoyberg/http-client/issues/71.

Since 0.3.8

proxyFromRequest :: ProxyOverride

Get the proxy settings from the Request itself.

Since 0.4.7

noProxy :: ProxyOverride

Never connect using a proxy, regardless of the proxy value in the Request.

Since 0.4.7

useProxy :: Proxy -> ProxyOverride

Use the given proxy settings, regardless of the proxy value in the Request.

Since 0.4.7

proxyEnvironment

Arguments

:: Maybe Proxy
fallback if no environment set

-> ProxyOverride

Get the proxy settings from the default environment variable (http_proxy for insecure, https_proxy for secure). If no variable is set, then fall back to the given value. Nothing is equivalent to noProxy, Just is equivalent to useProxy.

Since 0.4.7

proxyEnvironmentNamed

Arguments

:: Text
environment variable name

-> Maybe Proxy
fallback if no environment set

-> ProxyOverride

Same as proxyEnvironment, but instead of default environment variable names, allows you to set your own name.

Since 0.4.7

defaultProxy :: ProxyOverride

The default proxy settings for a manager. In particular: if the http_proxy (or https_proxy) environment variable is set, use it. Otherwise, use the values in the Request.

Since 0.4.7

dropProxyAuthSecure :: Request -> Request

Drop the Proxy-Authorization header from the request if we're using a secure proxy.

All types

type BodyReader = IO ByteString

An IO action that represents an incoming response body coming from the server. Data provided by this action has already been gunzipped and de-chunked, and respects any content-length headers present.

The action gets a single chunk of data from the response body, or an empty bytestring if no more data is available.

Since 0.4.0

data Connection

Constructors

Connection

Fields

connectionRead :: IO ByteString

If no more data, return empty.

connectionUnread :: ByteString -> IO ()

Return data to be read next time.

connectionWrite :: ByteString -> IO ()

Send data to server

connectionClose :: IO ()

Close connection. Any successive operation on the connection (exept closing) should fail with ConnectionClosed exception. It is allowed to close connection multiple times.

data StatusHeaders

Constructors

StatusHeaders Status HttpVersion RequestHeaders

data ConnectionClosed

Constructors

ConnectionClosed

data HttpException

Constructors

StatusCodeException Status ResponseHeaders CookieJar

InvalidUrlException String String

TooManyRedirects [Response ByteString]
List of encountered responses containing redirects in reverse chronological order; including last redirect, which triggered the exception and was not followed.

UnparseableRedirect (Response ByteString)
Response containing unparseable redirect.

TooManyRetries

HttpParserException String

HandshakeFailed

OverlongHeaders

ResponseTimeout

FailedConnectionException String Int
host/port

Note that in old versions of http-client and http-conduit, this exception would indicate a failed attempt to create a connection. However, since (at least) http-client 0.4, it indicates a timeout occurred while trying to establish the connection. For more information on this, see:

https://github.com/snoyberg/http-client/commit/b86b1cdd91e56ee33150433dedb32954d2082621#commitcomment-10718689

FailedConnectionException2 String Int Bool SomeException
host/port/secure

ExpectedBlankAfter100Continue

InvalidStatusLine ByteString

InvalidHeader ByteString

InternalIOException IOException

ProxyConnectException ByteString Int (Either ByteString HttpException)
host/port

NoResponseDataReceived

TlsException SomeException

TlsNotSupported

WrongRequestBodyStreamSize Word64 Word64
The request body provided did not match the expected size.

Provides the expected and actual size.

Since: 0.4.31

ResponseBodyTooShort Word64 Word64
Expected size/actual size.

Since 1.9.4

InvalidChunkHeaders
Since 1.9.4

IncompleteHeaders

InvalidDestinationHost ByteString

HttpZlibException ZlibException
Since 0.3

InvalidProxyEnvironmentVariable Text Text
Environment name and value

Since 0.4.7

ResponseLengthAndChunkingBothUsed
Detect a case where both the content-length header and transfer-encoding: chunked are used. Since 0.4.8.

Since 0.4.11 this exception isn't thrown anymore.

TlsExceptionHostPort SomeException ByteString Int
TLS exception, together with the host and port

Since: 0.4.24

data Cookie

Constructors

Cookie

Fields

cookie_name :: ByteString

cookie_value :: ByteString

cookie_expiry_time :: UTCTime

cookie_domain :: ByteString

cookie_path :: ByteString

cookie_creation_time :: UTCTime

cookie_last_access_time :: UTCTime

cookie_persistent :: Bool

cookie_host_only :: Bool

cookie_secure_only :: Bool

cookie_http_only :: Bool

newtype CookieJar

Constructors

CJ

Fields

expose :: [Cookie]

data Proxy

Define a HTTP proxy, consisting of a hostname and port number.

Constructors

Proxy

Fields

proxyHost :: ByteString

The host name of the HTTP proxy.

proxyPort :: Int

The port number of the HTTP proxy.

data RequestBody

When using one of the RequestBodyStream / RequestBodyStreamChunked constructors, you must ensure that the GivesPopper can be called multiple times. Usually this is not a problem.

The RequestBodyStreamChunked will send a chunked request body. Note that not all servers support this. Only use RequestBodyStreamChunked if you know the server you're sending to supports chunked request bodies.

Since 0.1.0

Constructors

RequestBodyLBS ByteString

RequestBodyBS ByteString

RequestBodyBuilder Int64 Builder

RequestBodyStream Int64 (GivesPopper ())

RequestBodyStreamChunked (GivesPopper ())

RequestBodyIO (IO RequestBody)
Allows creation of a RequestBody inside the IO monad, which is useful for making easier APIs (like setRequestBodyFile).

Since: 0.4.28

type Popper = IO ByteString

A function which generates successive chunks of a request body, provider a single empty bytestring when no more data is available.

Since 0.1.0

type NeedsPopper a = Popper -> IO a

A function which must be provided with a Popper.

Since 0.1.0

type GivesPopper a = NeedsPopper a -> IO a

A function which will provide a Popper to a NeedsPopper. This seemingly convoluted structure allows for creation of request bodies which allocate scarce resources in an exception safe manner.

Since 0.1.0

data Request

All information on how to connect to a host and what should be sent in the HTTP request.

If you simply wish to download from a URL, see parseRequest.

The constructor for this data type is not exposed. Instead, you should use either the defaultRequest value, or parseRequest to construct from a URL, and then use the records below to make modifications. This approach allows http-client to add configuration options without breaking backwards compatibility.

For example, to construct a POST request, you could do something like:

initReq <- parseRequest "http://www.example.com/path"
let req = initReq
            { method = "POST"
            }

For more information, please see http://www.yesodweb.com/book/settings-types.

Since 0.1.0

Constructors

Request

Fields

method :: Method

HTTP request method, eg GET, POST.

Since 0.1.0

secure :: Bool

Whether to use HTTPS (ie, SSL).

Since 0.1.0

host :: ByteString

Requested host name, used for both the IP address to connect to and the host request header.

Since 0.1.0

port :: Int

The port to connect to. Also used for generating the host request header.

Since 0.1.0

path :: ByteString

Everything from the host to the query string.

Since 0.1.0

queryString :: ByteString

Query string appended to the path.

Since 0.1.0

requestHeaders :: RequestHeaders

Custom HTTP request headers

The Content-Length and Transfer-Encoding headers are set automatically by this module, and shall not be added to requestHeaders.

If not provided by the user, Host will automatically be set based on the host and port fields.

Moreover, the Accept-Encoding header is set implicitly to gzip for convenience by default. This behaviour can be overridden if needed, by setting the header explicitly to a different value. In order to omit the Accept-Header altogether, set it to the empty string "". If you need an empty Accept-Header (i.e. requesting the identity encoding), set it to a non-empty white-space string, e.g. " ". See RFC 2616 section 14.3 for details about the semantics of the Accept-Header field. If you request a content-encoding not supported by this module, you will have to decode it yourself (see also the decompress field).

Note: Multiple header fields with the same field-name will result in multiple header fields being sent and therefore it's the responsibility of the client code to ensure that the rules from RFC 2616 section 4.2 are honoured.

Since 0.1.0

requestBody :: RequestBody

Request body to be sent to the server.

Since 0.1.0

proxy :: Maybe Proxy

Optional HTTP proxy.

Since 0.1.0

hostAddress :: Maybe HostAddress

Optional resolved host address. May not be used by all backends.

Since 0.1.0

rawBody :: Bool

If True, a chunked and/or gzipped body will not be decoded. Use with caution.

Since 0.1.0

decompress :: ByteString -> Bool

Predicate to specify whether gzipped data should be decompressed on the fly (see alwaysDecompress and browserDecompress). Argument is the mime type. Default: browserDecompress.

Since 0.1.0

redirectCount :: Int

How many redirects to follow when getting a resource. 0 means follow no redirects. Default value: 10.

Since 0.1.0

checkStatus :: Status -> ResponseHeaders -> CookieJar -> Maybe SomeException

Check the status code. Note that this will run after all redirects are performed. Default: return a StatusCodeException on non-2XX responses.

Since 0.1.0

responseTimeout :: Maybe Int

Number of microseconds to wait for a response. If Nothing, will wait indefinitely. Default: use managerResponseTimeout (which by default is 30 seconds).

Since 0.1.0

getConnectionWrapper :: Maybe Int -> HttpException -> IO (ConnRelease, Connection, ManagedConn) -> IO (Maybe Int, (ConnRelease, Connection, ManagedConn))

Wraps the calls for getting new connections. This can be useful for instituting some kind of timeouts. The first argument is the value of responseTimeout. Second argument is the exception to be thrown on failure.

Default: If responseTimeout is Nothing, does nothing. Otherwise, institutes timeout, and returns remaining time for responseTimeout.

Since 0.1.0

cookieJar :: Maybe CookieJar

A user-defined cookie jar. If Nothing, no cookie handling will take place, "Cookie" headers in requestHeaders will be sent raw, and responseCookieJar will be empty.

Since 0.1.0

requestVersion :: HttpVersion

HTTP version to send to server.

Default: HTTP 1.1

Since 0.4.3

onRequestBodyException :: SomeException -> IO ()

How to deal with exceptions thrown while sending the request.

Default: ignore IOExceptions, rethrow all other exceptions.

Since: 0.4.6

requestManagerOverride :: Maybe Manager

A Manager value that should override whatever Manager value was passed in to the HTTP request function manually. This is useful when dealing with implicit global managers, such as in Network.HTTP.Simple

Since: 0.4.28

data ConnReuse

Constructors

Reuse

DontReuse

type ConnRelease = ConnReuse -> IO ()

data ManagedConn

Constructors

Fresh

Reused

data Response body

A simple representation of the HTTP response.

Since 0.1.0

Constructors

Response

Fields

responseStatus :: Status

Status code of the response.

Since 0.1.0

responseVersion :: HttpVersion

HTTP version used by the server.

Since 0.1.0

responseHeaders :: ResponseHeaders

Response headers sent by the server.

Since 0.1.0

responseBody :: body

Response body sent by the server.

Since 0.1.0

responseCookieJar :: CookieJar

Cookies set on the client after interacting with the server. If cookies have been disabled by setting cookieJar to Nothing, then this will always be empty.

Since 0.1.0

responseClose' :: ResponseClose

Releases any resource held by this response. If the response body has not been fully read yet, doing so after this call will likely be impossible.

Since 0.1.0

newtype ResponseClose

Constructors

ResponseClose

Fields

runResponseClose :: IO ()

data Manager

Keeps track of open connections for keep-alive.

If possible, you should share a single Manager between multiple threads and requests.

Since 0.1.0

Constructors

Manager

Fields

mConns :: IORef ConnsMap

Nothing indicates that the manager is closed.

mConnsBaton :: MVar ()

Used to indicate to the reaper thread that it has some work to do. This must be filled every time a connection is returned to the manager. While redundant with the IORef above, this allows us to have the reaper thread fully blocked instead of running every 5 seconds when there are no connections to manage.

mMaxConns :: Int

This is a per-ConnKey value.

mResponseTimeout :: Maybe Int

Copied from managerResponseTimeout

mRawConnection :: Maybe HostAddress -> String -> Int -> IO Connection

mTlsConnection :: Maybe HostAddress -> String -> Int -> IO Connection

mTlsProxyConnection :: ByteString -> (Connection -> IO ()) -> String -> Maybe HostAddress -> String -> Int -> IO Connection

mRetryableException :: SomeException -> Bool

mWrapIOException :: forall a. IO a -> IO a

mIdleConnectionCount :: Int

mModifyRequest :: Request -> IO Request

mSetProxy :: Request -> Request

See managerProxy

class HasHttpManager a where

Minimal complete definition

getHttpManager

Methods

getHttpManager :: a -> Manager

data ConnsMap

Constructors

ManagerClosed

ManagerOpen !Int !(Map ConnKey (NonEmptyList Connection))

data ManagerSettings

Settings for a Manager. Please use the defaultManagerSettings function and then modify individual settings. For more information, see http://www.yesodweb.com/book/settings-types.

Since 0.1.0

Constructors

ManagerSettings

Fields

managerConnCount :: Int

Number of connections to a single host to keep alive. Default: 10.

Since 0.1.0

managerRawConnection :: IO (Maybe HostAddress -> String -> Int -> IO Connection)

Create an insecure connection.

Since 0.1.0 FIXME in the future, combine managerTlsConnection and managerTlsProxyConnection

managerTlsConnection :: IO (Maybe HostAddress -> String -> Int -> IO Connection)

Create a TLS connection. Default behavior: throw an exception that TLS is not supported.

Since 0.1.0

managerTlsProxyConnection :: IO (ByteString -> (Connection -> IO ()) -> String -> Maybe HostAddress -> String -> Int -> IO Connection)

Create a TLS proxy connection. Default behavior: throw an exception that TLS is not supported.

Since 0.2.2

managerResponseTimeout :: Maybe Int

Default timeout (in microseconds) to be applied to requests which do not provide a timeout value.

Default is 30 seconds

Since 0.1.0

managerRetryableException :: SomeException -> Bool

Exceptions for which we should retry our request if we were reusing an already open connection. In the case of IOExceptions, for example, we assume that the connection was closed on the server and therefore open a new one.

Since 0.1.0

managerWrapIOException :: forall a. IO a -> IO a

Action wrapped around all attempted Requests, usually used to wrap up exceptions in library-specific types.

Default: wrap all IOExceptions in the InternalIOException constructor.

Since 0.1.0

managerIdleConnectionCount :: Int

Total number of idle connection to keep open at a given time.

This limit helps deal with the case where you are making a large number of connections to different hosts. Without this limit, you could run out of file descriptors.

Default: 512

Since 0.3.7

managerModifyRequest :: Request -> IO Request

Perform the given modification to a Request before performing it.

Default: no modification

Since 0.4.4

managerProxyInsecure :: ProxyOverride

How HTTP proxy server settings should be discovered.

Default: respect the proxy value on the Request itself.

Since 0.4.7

managerProxySecure :: ProxyOverride

How HTTPS proxy server settings should be discovered.

Default: respect the proxy value on the Request itself.

Since 0.4.7

data NonEmptyList a

Constructors

One a UTCTime

Cons a Int UTCTime (NonEmptyList a)

data ConnHost

Hostname or resolved host address.

Constructors

HostName Text

HostAddress HostAddress

data ConnKey

ConnKey consists of a hostname, a port and a Bool specifying whether to use SSL.

Constructors

ConnKey ConnHost Int ByteString Int Bool

newtype ProxyOverride

How the HTTP proxy server settings should be discovered.

Since 0.4.7

Constructors

ProxyOverride

Fields

runProxyOverride :: Bool -> IO (Request -> Request)

data StreamFileStatus

Status of streaming a request body from a file.

Since 0.4.9

Constructors

StreamFileStatus

Fields

fileSize :: Int64

readSoFar :: Int64

thisChunkSize :: Int

Various utilities

hGetSome :: Handle -> Int -> IO ByteString

Like hGet, except that a shorter ByteString may be returned if there are not enough bytes immediately available to satisfy the whole request. hGetSome only blocks if there is no data available, and EOF has not yet been reached.

(<>) :: Monoid m => m -> m -> m infixr 5

readDec :: Integral i => String -> Maybe i

hasNoBody

Arguments

:: ByteString
request method

-> Int
status code

-> Bool

fromStrict :: ByteString -> ByteString

O(1) Convert a strict ByteString into a lazy ByteString.

timeout :: Int -> IO a -> IO (Maybe a)

Wrap an IO computation to time out and return Nothing in case no result is available within n microseconds (1/10^6 seconds). In case a result is available before the timeout expires, Just a is returned. A negative timeout interval means "wait indefinitely". When specifying long timeouts, be careful not to exceed maxBound :: Int.

The design of this combinator was guided by the objective that timeout n f should behave exactly the same as f as long as f doesn't time out. This means that f has the same myThreadId it would have without the timeout wrapper. Any exceptions f might throw cancel the timeout and propagate further up. It also possible for f to receive exceptions thrown to it by another thread.

A tricky implementation detail is the question of how to abort an IO computation. This combinator relies on asynchronous exceptions internally. The technique works very well for computations executing inside of the Haskell runtime system, but it doesn't work at all for non-Haskell code. Foreign function calls, for example, cannot be timed out with this combinator simply because an arbitrary C function cannot receive asynchronous exceptions. When timeout is used to wrap an FFI call that blocks, no timeout event can be delivered until the FFI call returns, which pretty much negates the purpose of the combinator. In practice, however, this limitation is less severe than it may sound. Standard I/O functions like hGetBuf, hPutBuf, Network.Socket.accept, or hWaitForInput appear to be blocking, but they really don't because the runtime system uses scheduling mechanisms like select(2) to perform asynchronous I/O, so it is possible to interrupt standard socket I/O or file I/O using this combinator.